feat : 미인증 회원의 접근에 대한 예외 처리, success

+++ src/main/java/nextstep/subway/ui/FavouriteController.java
@@ -5,6 +5,7 @@ import nextstep.subway.application.FavouriteService;
 import nextstep.subway.application.dto.FavouriteRequest;
 import nextstep.subway.application.dto.FavouriteResponse;
 import nextstep.subway.domain.member.MemberAdaptor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -23,20 +24,28 @@ public class FavouriteController {
     @PostMapping
     public ResponseEntity doFavourite(@AuthenticationPrincipal MemberAdaptor memberAdaptor,
                                       @RequestBody FavouriteRequest favouriteRequest) {
-        Long favouriteId = favouriteService.add(memberAdaptor.getId(), favouriteRequest);
-
-        return ResponseEntity.created(URI.create("/favourites/" + favouriteId)).build();
+        if (memberAdaptor.isEnabled()) {
+            Long favouriteId = favouriteService.add(memberAdaptor.getId(), favouriteRequest);
+            return ResponseEntity.created(URI.create("/favourites/" + favouriteId)).build();
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @GetMapping
     public ResponseEntity<List<FavouriteResponse>> getFavouriteList(@AuthenticationPrincipal MemberAdaptor memberAdaptor) {
-        return ResponseEntity.ok().body(favouriteService.findAll(memberAdaptor.getId()));
+        if (memberAdaptor.isEnabled()) {
+            return ResponseEntity.ok().body(favouriteService.findAll(memberAdaptor.getId()));
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @DeleteMapping("/{favouriteId}")
     public ResponseEntity<Void> deleteFavourite(@AuthenticationPrincipal MemberAdaptor memberAdaptor,
                                                 @PathVariable("favouriteId") Long favouriteId) {
-        favouriteService.delete(memberAdaptor.getId(), favouriteId);
-        return ResponseEntity.noContent().build();
+        if (memberAdaptor.isEnabled()) {
+            favouriteService.delete(memberAdaptor.getId(), favouriteId);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 }
